+++ src/main/java/kitchenpos/menu/service/MenuService.java
@@ -3,8 +3,8 @@ package kitchenpos.menu.service;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.menu.domain.entity.Menu;
-import kitchenpos.menu.domain.entity.MenuGroup;
-import kitchenpos.menu.domain.entity.MenuGroupRepository;
+import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroupRepository;
 import kitchenpos.menu.domain.entity.MenuProduct;
 import kitchenpos.menu.domain.entity.MenuRepository;
 import kitchenpos.menu.domain.value.Price;
