+++ src/main/java/kitchenpos/menugroup/domain/MenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain.entity;
+package kitchenpos.menugroup.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
