+++ src/test/java/kitchenpos/menu/service/MenuServiceTest.java
@@ -11,12 +11,12 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 import kitchenpos.menu.domain.entity.Menu;
-import kitchenpos.menu.domain.entity.MenuGroup;
-import kitchenpos.menu.domain.entity.MenuGroupRepository;
+import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroupRepository;
 import kitchenpos.menu.domain.entity.MenuProduct;
 import kitchenpos.menu.domain.entity.MenuRepository;
 import kitchenpos.menu.domain.value.Price;
-import kitchenpos.menu.dto.MenuGroupRequest;
+import kitchenpos.menugroup.dto.MenuGroupRequest;
 import kitchenpos.menu.dto.MenuProductRequest;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menu.dto.MenuResponse;
